# Remove unused sections for bpf_lxc from nodeport header