**Background**
One of the Kubernetes SLO is the Pod Startup Latency SLO. The current limit is the 99th percentile <= 5s. With the use of cilium Kubernetes is very close to this limit, sometimes even exceeding it. One of the major bottlenecks is eBPF program verification performed by kernel.

**Change description**
I went through the code to find kernel verification times for different sections, the most consuming ones were `tail_nodeport_nat_egress_ipv4` and `tail_nodeport_nat_egress_ipv6`. These sections are actually not used by bpf_lxc at all, but added and verified by kernel. So this change removes unused for bpf_lxc sections from nodeport header. After this change the total verification time for all programs for bpf_lxc reduced almost twice (from ~500ms to ~270ms). And with all the routine to load program, these changes contributed to the whole latency drop even more.

**Test scenario**
All tests were run on the cluster version 1.24.4-gke.300. The cilium master branch was used as a baseline for all the tests.

To test different configurations a GKE sandbox with custom Dataplane v2 images was used. The cluster was created with 100 nodes and DPv2 enabled. For Pod Startup Latency measurements a modified configuration of a GKE Scalability 100 nodes performance test was used.

In the beginning for each run 1 pod was created on each node with the same image as in Pod Startup Latency measurement, so image pull time is not affecting the resulting latency. Then 500 deployments with 5 pods each were created and waited until running to measure the pod startup latency.

**Results**

Result were aggregated from 20 separate test runs.

Percentile | Baseline | Optimized
--- | --- | ---
P50 | 3088 ms | 2584 ms
P90 | 3701 ms | 3169 ms
P99 | 4126 ms | 3573 ms

**Summary**
This optimization reduces the total latency by ~500ms and removes unnecessary sections loading to kernel.

```release-note
Remove unused sections for bpf_lxc from nodeport.h
```

Signed-off-by: Alex Katsman <alexkats@google.com>