@alexkats 2022-10-06
Hi, thanks for approval, I added a release note and resolved conflicts. I believe this change should have a `release-note/misc` label (based on what I see in the contributing guide), but I'm not able to change labels.