@joestringer 2022-10-18
This kind of change would typically not be a candidate for backporting based on the standard backporting criteria, as it is not a fix: https://docs.cilium.io/en/stable/contributing/release/backports/#backport-criteria 

That said, if we think the risk is low but benefit is high here then it may be worth backporting to 1.12.x. I know that there are occasionally user-reported issues around the ability of Cilium to create pods at a certain rate, and this could have a meaningful impact on those users. This seems at least reasonable to raise as a discussion point during the next [community meeting](https://docs.cilium.io/en/stable/community/community/#id1), I'll put it on the agenda.