@pchaigno 2022-10-19
Hi @alexkats,
We discussed this change and its backport in the community meeting today. It would help to have a bit more information on this bug (the broken SLO specifically) and how common it might be.

Specifically, could you share on which kernel(s) you made the measurements and with what Cilium configuration(s)?
> With the use of cilium Kubernetes is very close to this limit, sometimes even exceeding it.

In which conditions was it exceeding the limit? Your numbers show that it's always <5s, but I'm guessing some of the runs exceeded 5s? Is that still happening just less frequently with this patch?
> The cilium master branch was used as a baseline for all the tests. To test different configurations a GKE sandbox with custom Dataplane v2 images was used.

I'm also a bit confused by these two sentences. Did you test this with Cilium OSS or with GKE's patched Cilium version?