@alexkats 2022-10-24
Hi @pchaigno,

Thanks for the discussion.

I used GKE's patched Cilium version based on Cilium OSS master (that's why I stated that I used the master branch). The kernel version used was 5.10.109.

Actually, we try to aim for 3.5 seconds. We want to minimize the pod startup latency regression when changing kube-proxy based dataplane to Cilium on GKE. There is a k8s SLO - 99th percentile per cluster-day <= 5s, we have room to spend but we want to use as little as possible of it.

I performed tests with different number of nodes in the cluster, specifically 100, 500 and 5k. 
* With 100 nodes it didn't go beyond 5 seconds, but comes pretty close from time to time to the limit.
* When running with 500 nodes the average number for P99 is still within 5 seconds, but sometimes goes over and reaches around 5.5 seconds. 
* For 5k nodes the latency is at ~6 seconds on average, but goes even beyond it quite often.

With the change we go further away from the limit for 100 and 500 nodes, and become closer to 5 seconds for 5k nodes.