@pchaigno 2022-10-27
Hi @alexkats,
Thanks for the details!

It sounds like we still have a bug (SLO breakage) at scale. Could you open an issue so we can properly track that? We will also be able to refer to that issue as motivation when backporting the present pull request.

Did you collect Cilium metrics when running those benchmarks? It would be useful to understand where the additional pod creation latency is coming from (e.g., slow kube-apiserver at scale).