@alexkats 2022-11-07
Hi @pchaigno ,

I created the issue #22023.

Regarding metrics, I didn't really take a look into them yet, but I plan to do it when running new tests.